backend/requirements.txt
# ── Core scanner deps ────────────────────────────────────────────────────────
yara-python>=4.3.1
gitpython>=3.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
attrs==25.4.0
boltons==21.0.0
bracex==2.6
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
click-option-group==0.5.9
colorama==0.4.6
exceptiongroup==1.2.2
face==24.0.0
fastapi==0.124.2
glom==22.1.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0

protobuf==6.33.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
referencing==0.37.0
requests==2.32.5
rich==13.5.3
rpds-py==0.30.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14

sse-starlette==3.0.3
starlette==0.50.0
tomli==2.0.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.2
uvicorn==0.38.0
wcmatch==8.5.2
wrapt==1.17.3
zipp==3.23.0
